requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytesseract
pillow
transformers
torch
sentence-transformers
chromadb
sentencepiece